server-app/requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
django-cors-headers==3.8.0
pytz==2021.1
sqlparse==0.4.2
psycopg2-binary==2.9.1
django-axes==5.25.0
django-session-timeout==0.1.0
pip==21.1.3

django-nested-admin~=3.3.3
PyJWT~=1.7.1
cryptography~=3.4.7
keyring~=18.0.1
colorama~=0.4.3
PyYAML~=5.3.1
Jinja2~=3.0.1
numpy~=1.19.4
setuptools~=45.2.0
requests~=2.22.0
Markdown~=3.3.3
Pygments~=2.7.2

python-telegram-bot==13.7
